refactor(pkg): move kafka helpers to pkg/kafka
