+++ pkg/kafka/kafka_util.go
