+++ pkg/kafka/kafka_util_test.go
